+++ include/opcode_dict.h
@@ -39,8 +39,8 @@ opcode_stub_t opcode_stub_by_code(opcode_dict_t, word_t);
 word_t opcode_code_by_stub(opcode_dict_t, opcode_stub_t);
 const char* opcode_name_by_stub(opcode_dict_t, opcode_stub_t);
 
-int opcode_dict_link_stubs(opcode_dict_t target, opcode_dict_t src);
-int opcode_dict_resolve_stubs(opcode_dict_t src);
+long opcode_dict_link_stubs(opcode_dict_t target, opcode_dict_t src);
+long opcode_dict_resolve_stubs(opcode_dict_t src);
 
 
 void opcode_dict_serialize(opcode_dict_t od, writer_t);
